♻️Add type to MIN-files and TRN-files

+++ Source/diablo.cpp
@@ -1505,11 +1505,11 @@ void LoadLvlGFX()
 		if (gbIsHellfire) {
 			pDungeonCels = LoadFileInMem<BYTE>("NLevels\\TownData\\Town.CEL");
 			pMegaTiles = LoadFileInMem<MegaTile>("NLevels\\TownData\\Town.TIL");
-			pLevelPieces = LoadFileInMem<BYTE>("NLevels\\TownData\\Town.MIN");
+			pLevelPieces = LoadFileInMem<uint16_t>("NLevels\\TownData\\Town.MIN");
 		} else {
 			pDungeonCels = LoadFileInMem<BYTE>("Levels\\TownData\\Town.CEL");
 			pMegaTiles = LoadFileInMem<MegaTile>("Levels\\TownData\\Town.TIL");
-			pLevelPieces = LoadFileInMem<BYTE>("Levels\\TownData\\Town.MIN");
+			pLevelPieces = LoadFileInMem<uint16_t>("Levels\\TownData\\Town.MIN");
 		}
 		pSpecialCels = LoadCel("Levels\\TownData\\TownS.CEL", SpecialCelWidth);
 		break;
@@ -1517,37 +1517,37 @@ void LoadLvlGFX()
 		if (currlevel < 21) {
 			pDungeonCels = LoadFileInMem<BYTE>("Levels\\L1Data\\L1.CEL");
 			pMegaTiles = LoadFileInMem<MegaTile>("Levels\\L1Data\\L1.TIL");
-			pLevelPieces = LoadFileInMem<BYTE>("Levels\\L1Data\\L1.MIN");
+			pLevelPieces = LoadFileInMem<uint16_t>("Levels\\L1Data\\L1.MIN");
 			pSpecialCels = LoadCel("Levels\\L1Data\\L1S.CEL", SpecialCelWidth);
 		} else {
 			pDungeonCels = LoadFileInMem<BYTE>("NLevels\\L5Data\\L5.CEL");
 			pMegaTiles = LoadFileInMem<MegaTile>("NLevels\\L5Data\\L5.TIL");
-			pLevelPieces = LoadFileInMem<BYTE>("NLevels\\L5Data\\L5.MIN");
+			pLevelPieces = LoadFileInMem<uint16_t>("NLevels\\L5Data\\L5.MIN");
 			pSpecialCels = LoadCel("NLevels\\L5Data\\L5S.CEL", SpecialCelWidth);
 		}
 		break;
 	case DTYPE_CATACOMBS:
 		pDungeonCels = LoadFileInMem<BYTE>("Levels\\L2Data\\L2.CEL");
 		pMegaTiles = LoadFileInMem<MegaTile>("Levels\\L2Data\\L2.TIL");
-		pLevelPieces = LoadFileInMem<BYTE>("Levels\\L2Data\\L2.MIN");
+		pLevelPieces = LoadFileInMem<uint16_t>("Levels\\L2Data\\L2.MIN");
 		pSpecialCels = LoadCel("Levels\\L2Data\\L2S.CEL", SpecialCelWidth);
 		break;
 	case DTYPE_CAVES:
 		if (currlevel < 17) {
 			pDungeonCels = LoadFileInMem<BYTE>("Levels\\L3Data\\L3.CEL");
 			pMegaTiles = LoadFileInMem<MegaTile>("Levels\\L3Data\\L3.TIL");
-			pLevelPieces = LoadFileInMem<BYTE>("Levels\\L3Data\\L3.MIN");
+			pLevelPieces = LoadFileInMem<uint16_t>("Levels\\L3Data\\L3.MIN");
 		} else {
 			pDungeonCels = LoadFileInMem<BYTE>("NLevels\\L6Data\\L6.CEL");
 			pMegaTiles = LoadFileInMem<MegaTile>("NLevels\\L6Data\\L6.TIL");
-			pLevelPieces = LoadFileInMem<BYTE>("NLevels\\L6Data\\L6.MIN");
+			pLevelPieces = LoadFileInMem<uint16_t>("NLevels\\L6Data\\L6.MIN");
 		}
 		pSpecialCels = LoadCel("Levels\\L1Data\\L1S.CEL", SpecialCelWidth);
 		break;
 	case DTYPE_HELL:
 		pDungeonCels = LoadFileInMem<BYTE>("Levels\\L4Data\\L4.CEL");
 		pMegaTiles = LoadFileInMem<MegaTile>("Levels\\L4Data\\L4.TIL");
-		pLevelPieces = LoadFileInMem<BYTE>("Levels\\L4Data\\L4.MIN");
+		pLevelPieces = LoadFileInMem<uint16_t>("Levels\\L4Data\\L4.MIN");
 		pSpecialCels = LoadCel("Levels\\L2Data\\L2S.CEL", SpecialCelWidth);
 		break;
 	default:
